changed structuring, moved state out to App
