+++ src/constants/actionTypes.js
@@ -1,4 +1,4 @@
 // boilerplate template - needs to be updated with more suitable actionTypes
 
-export const GET_EC2 = 'GET_EC2';
+export const GET_AWS_INSTANCES = 'GET_AWS_INSTANCES';
 export const NODE_DETAILS = 'NODE_DETAILS';
