+++ src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 
-import App from "./components/App";
+import App from "./containers/App";
 import { Provider } from "react-redux";
 import store from "./store";
 import './styles/App.scss'
